Add timeouts for the all tests in the `dropwizard-jersey`module
Recently, the tests for Jersey resources were rather flaky and failed on Travis CI.
Unfortunately, we can't get any debug information about it, because they the
test suite is killed by the CI environment because it becomes unresponsive.
It would be great to have timeouts for each tests, so even in case we have errors,
the whole test suite won't fail.

This is achieved by introducing a new class `AbstractJerseyTest` which contains
common features of the all tests for Jersey resources. Now the tests are inherited
from it and perform HTTP requests with timeouts.