Add more Liquibase tests (#1915)
* Add tests for DbCalculatechecksumCommand

* Add tests for DbClearChecksumsCommand

* Add tests for DbDropAllCommand

* Add tests for DbFastForward command

* Add tests for DbGenerateDocs command

* Add tests for DbLocksCommand

* Add tests for DbPrepareRollbackCommand

* Add tests for DbRollbackCommand

* Add tests for DbTagCommand

Also make the `openLiquibase` method open, so we can get a `Liquibase` object
in the tests.

* Add tests for DbTestCommand

* Use the UTF_8 constant in migration tests

* Rename `migrations-test.xml` to `migrations-ddl.xml`

`migration-test` is a rather ambiguous name.

* Add tests for DbCommand

* Add a test for migrations from the filesystem

Verify if we pass the absolute path to file on the filesystem instead
the relative path, migrations still work.

* Mark all migration tests as not-thread safe

So they don't blow up and run in one thread when Maven run the tests in parallel.

* Print the help page in UTF-8 in migration tests

To make sure we read and write in the same encoding. Moreover, error-prone
gives a warning about writing without an explicit encoding.

* Run all migration tests in memory

We can run the database in memory with the `close_delay=-1` parameter
instead of the file mode. That parameter means that the database should
be closed only when the JVM exits, not when the JDBC connection is
closed. Hence, the changes performed by one command are visible to the
others in the same JVM, which is exactly what we need.