Update to Jackson 2.9
Jackson 2.9 greatly simplifies exception handling with the new exception,
`InvalidDefinitionException`.

This PR technically breaks behavior for custom deserializers, but I
believe it is now the correct. If a deserializer fails to to deserialize
input then a 400 should be returned as it denotes unexpected input. A 500
should only be returned if a deserializer encountered an error unrelated
to the input (if it read from a non-existant file or db...(not saying it's
a good design)). Added a test case to assert this.