Move UnitOfWork state into a publicly available thread container
UnitOfWorkContext uses ThreadLocals to keep track of a UnitOfWork context variables tied to a thread (e.g., a HTTP request). This is a simple map similar to org.hibernate.context.internal.ManagedSessionContext. Since ManagedSessionContext is meant to be used for purposes internal to Hibernate, UnitOfWorkContext exists to be public to Dropwizard clients.

In order to avoid tracking context variables in multiple places, the `unitOfWork` variable was also moved to this static context.